Move ProcessStatusUpdateRecord to core module
