+++ java/compaction-core/src/test/java/sleeper/compaction/job/CompactionJobStatusesBuilderTest.java
@@ -19,12 +19,12 @@ import org.junit.Test;
 import sleeper.compaction.job.status.CompactionJobCreatedStatus;
 import sleeper.compaction.job.status.CompactionJobStatus;
 import sleeper.compaction.job.status.CompactionJobStatusesBuilder;
-import sleeper.compaction.job.status.ProcessStatusUpdateRecord;
 import sleeper.core.record.process.RecordsProcessed;
 import sleeper.core.record.process.RecordsProcessedSummary;
 import sleeper.core.status.ProcessFinishedStatus;
 import sleeper.core.status.ProcessRun;
 import sleeper.core.status.ProcessStartedStatus;
+import sleeper.core.status.ProcessStatusUpdateRecord;
 
 import java.time.Instant;
 import java.util.Arrays;
