+++ java/core/src/main/java/sleeper/core/status/ProcessStatusUpdateRecord.java
@@ -13,9 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.job.status;
-
-import sleeper.core.status.ProcessStatusUpdate;
+package sleeper.core.status;
 
 import java.time.Instant;
 import java.util.Objects;
